+++ mlos_bench/mlos_bench/tunables/tunable.py
@@ -7,6 +7,12 @@ Definitions for :py:class:`~.Tunable` parameters.
 
 Tunable parameters are one of the core building blocks of the :py:mod:`mlos_bench`
 framework.
+Together with :py:class:`~mlos_bench.tunables.tunable_groups.TunableGroups`, they
+provide a description of a configuration parameter space for a benchmark or an
+autotuning optimization task.
+
+Some details about the configuration of an individual :py:class:`~.Tunable`
+parameter are available in the Examples docstrings below.
 
 See Also
 --------
@@ -15,223 +21,26 @@ See Also
 """
 import copy
 import logging
-from collections.abc import Iterable, Sequence
-from typing import Any, Literal, TypedDict
+from collections.abc import Iterable
+from typing import Any
 
 import numpy as np
 
+from mlos_bench.tunables.tunable_types import (
+    TunableValueTypeName,
+    TunableValueType,
+    TunableValue,
+    DistributionName,
+    tunable_dict_from_dict,
+    TUNABLE_DTYPE,
+)
 from mlos_bench.util import nullable
 
 _LOG = logging.getLogger(__name__)
 
-TunableValue = int | float | str | None
-"""A Tunable parameter value type alias."""
-
-TunableValueType = type[int] | type[float] | type[str]
-"""A :py:class:`TypeAlias` for :py:class:`.Tunable` value :py:attr:`~.Tunable.type`.
-
-See Also
---------
-:py:attr:`.Tunable.dtype` : Example of accepted types.
-"""
-
-TunableValueTypeTuple = (int, float, str, type(None))
-"""
-Tunable value type tuple.
-
-For checking with :py:func:`isinstance`.
-"""
-
-TunableValueTypeName = Literal["int", "float", "categorical"]
-"""
-The accepted string names of a :py:class:`.Tunable` value :py:attr:`~.Tunable.type`.
-
-See Also
---------
-:py:attr:`.Tunable.type` : Example of accepted type names.
-"""
-# pydoctest doesn't scan variable docstrings so we put the examples elsewhere.
-
-TunableValuesDict = dict[str, TunableValue]
-"""Tunable values dictionary type."""
-
-DistributionName = Literal["uniform", "normal", "beta"]
-"""
-The :py:class:`.Tunable` value :py:attr:`~.Tunable.distribution` type names.
-
-See Also
---------
-:py:attr:`.Tunable.distribution` : Example of accepted distribution names.
-"""
-# pydoctest doesn't scan variable docstrings so we put the examples elsewhere.
-
-
-class DistributionDictOpt(TypedDict, total=False):
-    """
-    A TypedDict for a :py:class:`.Tunable` parameter's optional
-    :py:attr:`~.Tunable.distribution_params` config.
-
-    Mostly used by type checking. These are the types expected to be received from
-    the json config.
-
-    Notes
-    -----
-    :py:class:`.DistributionDict` contains the required fields for the
-    :py:class:`.Tunable`'s :py:attr:`~.Tunable.distribution` parameter.
-
-    See Also
-    --------
-    :py:attr:`.Tunable.distribution_params` : Example of distribution params.
-    """
-
-    params: dict[str, float] | None
-
-
-class DistributionDict(DistributionDictOpt):
-    """
-    A TypedDict for a :py:class:`.Tunable` parameter's required ``distribution`` 's
-    config parameters.
-
-    Mostly used by type checking. These are the types expected to be received from the
-    json config.
-
-    Examples
-    --------
-    >>> # Example values of the DistributionName
-    >>> from mlos_bench.tunables.tunable import DistributionName
-    >>> DistributionName
-    typing.Literal['uniform', 'normal', 'beta']
-
-    >>> # Example values of the DistributionDict
-    >>> from mlos_bench.tunables.tunable import DistributionDict
-    >>> DistributionDict({'type': 'uniform'})
-    {'type': 'uniform'}
-    >>> DistributionDict({'type': 'normal', 'params': {'mu': 0.0, 'sigma': 1.0}})
-    {'type': 'normal', 'params': {'mu': 0.0, 'sigma': 1.0}}
-    >>> DistributionDict({'type': 'beta', 'params': {'alpha': 1.0, 'beta': 1.0}})
-    {'type': 'beta', 'params': {'alpha': 1.0, 'beta': 1.0}}
-
-    See Also
-    --------
-    :py:attr:`.Tunable.distribution` : Example of distributions.
-    :py:attr:`.Tunable.distribution_params` : Example of distribution params.
-    """
-
-    type: DistributionName
-
-
-class TunableDictOpt(TypedDict, total=False):
-    """
-    A TypedDict for a :py:class:`.Tunable` parameter's optional config parameters.
-
-    Mostly used for type checking. These are the types expected to be received from
-    the json config.
-
-    Notes
-    -----
-    :py:class:`TunableDict` contains the required fields for the
-    :py:class:`.Tunable` parameter.
-    """
-
-    # Optional fields
-    description: str | None
-    values: list[str | None] | None
-    range: Sequence[int] | Sequence[float] | None
-    quantization_bins: int | None
-    log: bool | None
-    distribution: DistributionDict | None
-    special: list[int] | list[float] | None
-    values_weights: list[float] | None
-    special_weights: list[float] | None
-    range_weight: float | None
-    meta: dict[str, Any]
-
-
-class TunableDict(TunableDictOpt):
-    """
-    A TypedDict for a :py:class:`.Tunable` parameter's required config parameters.
-
-    Mostly used for type checking. These are the types expected to be received from
-    the json config.
-
-    Examples
-    --------
-    >>> # Example values of the TunableDict
-    >>> from mlos_bench.tunables.tunable import TunableDict
-    >>> TunableDict({'type': 'int', 'default': 0, 'range': [0, 10]})
-    {'type': 'int', 'default': 0, 'range': [0, 10]}
-
-    >>> # Example values of the TunableDict with optional fields
-    >>> TunableDict({'type': 'categorical', 'default': 'a', 'values': ['a', 'b']})
-    {'type': 'categorical', 'default': 'a', 'values': ['a', 'b']}
-    """
-
-    # Required fields
-    type: TunableValueTypeName
-    default: TunableValue
-
-
-def tunable_dict_from_dict(config: dict[str, Any]) -> TunableDict:
-    """
-    Creates a TunableDict from a regular dict.
-
-    Parameters
-    ----------
-    config : dict[str, Any]
-        A regular dict that represents a TunableDict.
-
-    Returns
-    -------
-    TunableDict
-
-    Examples
-    --------
-    >>> # Example values of the TunableDict
-    >>> import json5 as json
-    >>> from mlos_bench.tunables.tunable import tunable_dict_from_dict
-    >>> config = json.loads("{'type': 'int', 'default': 0, 'range': [0, 10]}")
-    >>> config
-    {'type': 'int', 'default': 0, 'range': [0, 10]}
-    >>> typed_dict = tunable_dict_from_dict(config)
-    >>> typed_dict
-    {'type': 'int', 'description': None, 'default': 0, 'values': None, 'range': [0, 10], 'quantization_bins': None, 'log': None, 'distribution': None, 'special': None, 'values_weights': None, 'special_weights': None, 'range_weight': None, 'meta': {}}
-    """  # pylint: disable=line-too-long # noqa: E501
-    _type = config.get("type")
-    if _type not in Tunable.DTYPE:
-        raise ValueError(f"Invalid parameter type: {_type}")
-    _meta = config.get("meta", {})
-    return TunableDict(
-        type=_type,
-        description=config.get("description"),
-        default=config.get("default"),
-        values=config.get("values"),
-        range=config.get("range"),
-        quantization_bins=config.get("quantization_bins"),
-        log=config.get("log"),
-        distribution=config.get("distribution"),
-        special=config.get("special"),
-        values_weights=config.get("values_weights"),
-        special_weights=config.get("special_weights"),
-        range_weight=config.get("range_weight"),
-        meta=_meta,
-    )
-
 
 class Tunable:  # pylint: disable=too-many-instance-attributes,too-many-public-methods
     """A Tunable parameter definition and its current value."""
-
-    DTYPE: dict[TunableValueTypeName, TunableValueType] = {
-        "int": int,
-        "float": float,
-        "categorical": str,
-    }
-    """
-    Maps Tunable types to their corresponding Python data types by name.
-
-    See Also
-    --------
-    :py:attr:`.Tunable.dtype` : Example of type mappings.
-    """
 
     def __init__(self, name: str, config: dict):
         """
@@ -254,7 +63,7 @@ class Tunable:  # pylint: disable=too-many-instance-attributes,too-many-public-m
             raise ValueError(f"Invalid name of the tunable: {name}")
         self._name = name
         self._type: TunableValueTypeName = t_config["type"]  # required
-        if self._type not in self.DTYPE:
+        if self._type not in TUNABLE_DTYPE:
             raise ValueError(f"Invalid parameter type: {self._type}")
         self._description = t_config.get("description")
         self._default = t_config["default"]
@@ -700,7 +509,7 @@ class Tunable:  # pylint: disable=too-many-instance-attributes,too-many-public-m
         >>> Tunable.DTYPE
         {'int': <class 'int'>, 'float': <class 'float'>, 'categorical': <class 'str'>}
         """
-        return self.DTYPE[self._type]
+        return TUNABLE_DTYPE[self._type]
 
     @property
     def is_categorical(self) -> bool:
